Organize experimental modules and document architecture
