+++ memory/cortex.py
