+++ memory/emotional.py
