+++ memory/mental.py
