+++ memory/spiral_cortex.py
