+++ tests/test_cortex_memory.py
@@ -5,7 +5,7 @@ from pathlib import Path
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 
-import cortex_memory
+from memory import cortex as cortex_memory
 
 
 class DummyNode:
