+++ tests/test_memory_spiritual.py
@@ -1,17 +1,12 @@
 """Tests for the spiritual memory ontology database."""
 
-from memory_spiritual import (
-    get_connection,
-    map_to_symbol,
-    get_event_symbol,
-    lookup_symbol_history,
-)
+from memory import spiritual as ms
 
 
 def test_connection_initializes_schema():
     """An in-memory connection creates the schema automatically."""
 
-    conn = get_connection(":memory:")
+    conn = ms.get_connection(":memory:")
     cur = conn.execute(
         "SELECT name FROM sqlite_master WHERE type='table' AND name='event_symbol'"
     )
@@ -23,8 +18,8 @@ def test_event_symbol_roundtrip():
     """Storing and retrieving a symbol works with in-memory DB."""
 
     conn = get_connection(":memory:")
-    map_to_symbol(("rain", "☔"), conn=conn)
-    assert get_event_symbol("rain", conn=conn) == "☔"
-    assert lookup_symbol_history("☔", conn=conn) == ["rain"]
+    ms.map_to_symbol(("rain", "☔"), conn=conn)
+    assert ms.get_event_symbol("rain", conn=conn) == "☔"
+    assert ms.lookup_symbol_history("☔", conn=conn) == ["rain"]
     conn.close()
